requirements.txt
python-dotenv
yt-dlp
transformers
gradio
openai

